+++ src/main/java/org/example/model/Product.java
@@ -3,18 +3,18 @@ package org.example.model;
 
 public class Product extends EntityByString{
 
-    private Float price;
+    private Double price;
     private String origin;
 
-    public Product(String name, Float price, String origin){
+    public Product(String name, Double price, String origin){
         super(name);
         this.price=price;
         this.origin=origin;
     }
 
     public String getOrigin(){return this.origin;}
-    public Float getPrice(){return this.price;}
+    public Double getPrice(){return this.price;}
 
     public void setOrigin(String origin){this.origin=origin;}
-    public void setPrice(Float price){this.price=price;}
+    public void setPrice(Double price){this.price=price;}
 }
